extract util function to avoid side effects during test
